fix: rename film card
