+++ project/src/components/small-film-card/small-film-card.tsx
