+++ project/src/pages/movie-page-reviews/movie-page-reviews.tsx
@@ -1,4 +1,4 @@
-import SmallFilmCard from "../../components/smallFilmCard/smallFilmCard";
+import SmallFilmCard from "../../components/small-film-card/small-film-card";
 
 function MoviePageReviews(): JSX.Element {
   return (
