+++ project/src/pages/my-list/my-list.tsx
@@ -1,4 +1,4 @@
-import SmallFilmCard from '../../components/smallFilmCard/smallFilmCard';
+import SmallFilmCard from '../../components/small-film-card/small-film-card';
 
 function MyList(): JSX.Element {
   return (
